Support SSL on Ruby 1.8.7 / JRuby 1.6
Under older versions of Ruby, SSL needed to be turned on after creating the
HTTP object, but before making a request, as a separate method call; and you
must require 'net/https'.

In addition to the above, this brings more consistency under which
circumstances the SSL and proxy configuration variables are applied when
using server_ip instead of server_url to configure the location of the
Jenkins server.