Downgrade rspec to 2.13.0
In the recent version of rspec, the `{}.not_to raise_error(SpecificErrorClass)`
expectation is deprecated. In the Client class the API related calls are just
checked to see if they exist and accept the specified number of arguments.
Since the HTTP requests can't be simply mocked out, locking the rspec gem to a
known working version until further changes are made.