Make node functional spec pending
The node creation mainly uses the private key file to create nodes that use SSH.
The recent version of Jenkins got rid of the private key option and uses the credentials
plugin that creates SSH and username/password credentials. But the credentials plugin
doesn't have a REST interface that lets create credentials or obtain created credentials.
So node creation support in the API client requires that the credentials exposes the
functionality through the REST interface.

That's the story for disabling this test!